플레이어 스텟 스크립트 추가
