+++ Assets/scripts/Model/playerStat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2a0a05e03ccf69488a5ee070d8a0ad8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
